agent-python/pyproject.toml
[project]
name = "agent-python"
version = "0.1.0"
authors = [
    { name = "Mohid Yousaf", email = "mohid.yousaf@whizzbridge.com" }
]
requires-python = ">=3.13"
dependencies = [
    "email-validator>=2.3.0",
    "livekit-agents[openai]~=1.2",
    "livekit-plugins-noise-cancellation~=0.2",
    "python-dotenv>=1.1.1",
]

[build-system]
requires = ["uv_build>=0.9.0,<0.10.0"]
build-backend = "uv_build"
